+++ app/models/__init__.py
@@ -1,3 +1,6 @@
 from .db import db
 from .user import User
 from .reflection import Reflection
+from .comment import Comment
+from .jurisdiction import Jurisdiction
+from .pd import Pd
