+++ app/models/comment.py
@@ -1,0 +1,20 @@
+from .db import db
+
+
+class Comment(db.Model):
+    __tablename__ = 'comments'
+
+    id = db.Column(db.Integer, primary_key=True)
+    comment = db.Column(db.Text, nullable=False)
+    user_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
+    user = db.relationship("User", back_populates="comments")
+    reflection_id = db.Column(db.Integer, db.ForeignKey("reflections.id"), nullable=False)
+    post = db.relationship("Reflection", back_populates="comments")
+
+    def to_dict(self):
+        return {
+            "id": self.id,
+            "comment": self.comment,
+            "user_id": self.user_id,
+            "reflection_id": self.reflection_id
+        }
