--- app/models/comments.py
@@ -1,20 +1,0 @@
-from .db import db
-
-
-class Comment(db.model):
-    __tablename__ = 'comment'
-
-    id = db.Column(db.Integer, primary_key=True)
-    comment = db.Column(db.text, nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
-    user = db.relationship("User", back_populates="comments")
-    post_id = db.Column(db.Integer, db.ForeignKey("post.id"), nullable=False)
-    post = db.relationship("Post", back_populates="comments")
-
-    def to_dict(self):
-        return {
-            "id": self.id,
-            "comment": self.comment,
-            "user_id": self.user_id,
-            "post_id": self.post_id
-        }
