+++ app/models/pd.py
@@ -3,12 +3,12 @@ from .db import db
 
 
 class Pd(db.Model):
-    __tablename__ = 'pd'
+    __tablename__ = 'pds'
 
     id = db.Column(db.Integer, primary_key=True)
     department_name = db.Column(db.String(100), nullable=False)
-    jurisdictions = relationship("Jurisdiction", back_populates="pd")
-    reflections = relationship("Reflection", back_populates="pd")
+    jurisdictions = db.relationship("Jurisdiction", back_populates="pd")
+    reflections = db.relationship("Reflection", back_populates="pd")
 
     def to_dict(self):
         return {
